+++ Tests.h
@@ -1,0 +1,22 @@
+#pragma once
+#include <SFML/Graphics.hpp>
+#include "iostream"
+
+#include "MathFunction.h"
+#include "Surface.h"
+#include "Object.h"
+#include "RigidBody.h"
+#include "Ship.h"
+#include "Geom/Geometric.h"
+#include "Menu.h"
+
+using namespace sf;
+
+void Test1();
+void Test2();
+void Test3();
+void Test4();
+void Test5();
+void test_B2();
+void test_B1();
+void test_B3();
